- added littercoin init scripts - re-org scripts into folder for clarity - switched to optimzie=true for production optimizations
