+++ littercoin/src/threadToken.hl
@@ -1,0 +1,26 @@
+minting threadToken
+
+enum Redeemer { 
+    Init 
+}
+
+const TX_ID: ByteArray = #9e27c544200815e4277dad55f198ab2d2d67c746ff19a57cd4554e7b339a39a5
+const txId: TxId = TxId::new(TX_ID)
+const outputId: TxOutputId = TxOutputId::new(txId, 0)
+
+func main(ctx: ScriptContext) -> Bool {
+    tx: Tx = ctx.tx;
+    mph: MintingPolicyHash = ctx.get_current_minting_policy_hash();
+
+    tt_assetclass: AssetClass = AssetClass::new(
+        mph, 
+        "Thread Token Littercoin".encode_utf8()
+    );
+
+    value_minted: Value = tx.minted;
+    (value_minted == Value::new(tt_assetclass, 1)).trace("TT1: ") &&
+    tx.inputs.any((input: TxInput) -> Bool {
+                        (input.output_id == outputId).trace("TT2: ")
+                        }
+    )
+}
