+++ public/contracts/lcValidator.hl
@@ -22,7 +22,7 @@ const minAda : Int = 2000000
 const minAdaVal: Value = Value::lovelace(minAda)
 
 // Define thread token value
-const TT_MPH: ByteArray = #d21d0071d4bfc20583c142e755c191d61583d40e8f61a9ebe1d1ea62
+const TT_MPH: ByteArray = #a7add6eea42ec499442e590f1990203a0c66d9b759dc00fb10080200
 const ttMph: MintingPolicyHash = MintingPolicyHash::new(TT_MPH)
 const ttAssetclass: AssetClass = AssetClass::new(
         ttMph, 
@@ -31,7 +31,7 @@ const ttAssetclass: AssetClass = AssetClass::new(
 const ttVal : Value = Value::new(ttAssetclass, 1)
 
 // Define the mph of the littercoin minting policy
-const LC_MPH: ByteArray = #50963a51dc23b0c588812877bf2c6b1c53c4bac492df73a6e64993c9
+const LC_MPH: ByteArray = #108985b19877e4bd4300d0e18a90239471ac475f54ded96a7f014efd
 const lcMph: MintingPolicyHash = MintingPolicyHash::new(LC_MPH)
 const lcAssetClass: AssetClass = AssetClass::new(
         lcMph, 
@@ -39,7 +39,7 @@ const lcAssetClass: AssetClass = AssetClass::new(
     )
 
 // Define the mph of the donation rewards token minting policy
-const REWARDS_MPH: ByteArray = #0e30451846a76a70678796e7ff62516dc5196865e724b8a41b6c06c7
+const REWARDS_MPH: ByteArray = #5ae137d386575082c5649c8e057b1fae349f34246bd0ded4d5fe6eaf
 const rewardsMph: MintingPolicyHash = MintingPolicyHash::new(REWARDS_MPH)
 const rewardsAssetClass: AssetClass = AssetClass::new(
         rewardsMph, 
@@ -173,12 +173,4 @@ func main(datum: Datum, redeemer: Redeemer, ctx: ScriptContext) -> Bool {
             }
         }
     }
-}
-
-// Datum field order is important. Initialize
-// with 2 littercoin because the users can't drain
-// the littercoin smart contract lower than 2 Ada
-const LC_DATUM = Datum {
-    adaAmount : 2000000,
-    lcAmount : 2 
 }
